backend/requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
pydantic-settings==2.11.0
supabase==2.21.1
python-jose==3.5.0
python-multipart
python-dotenv==1.0.1